pyproject.toml
[project]
name = "teaicher"
version = "0.1.0"
description = "An AI storytelling tool"
authors = [{ name = "Morgan Saunier" }]
dependencies = [
  "gradio==5.23.1",
  "huggingface_hub==0.28.1",
  "python-vlc",
  "google-api-python-client",
  "spotipy",
  "openai",
  "isodate",
  "elevenlabs",
  "TTS",
  "torch",
  "Flask",
  "python-dotenv",
  "mutagen"
]

[project.optional-dependencies]
dev = ["python-dotenv"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
